Make Mangaflix more readable
added a few breaks on the longer lines.
The original script used the pattern 
```bash
LABEL=$(action)
```
to perform many actions. I find this hard to parse when compared to 
```sh
# LABEL
action
```
which also avoids spawning a whole new shell for each command.

I also replaced instances of `$(...)` to spawn subcommands with `\`...\``, which will work on `dash` and other posix compatible shells.